+++ packages/rs-sdk/src/platform/dashpay/contact_request.rs
@@ -5,11 +5,13 @@
 use crate::platform::transition::put_document::PutDocument;
 use crate::platform::Document;
 use crate::{Error, Sdk};
+use dash_platform_queries::dashpay::{
+    build_contact_request_document, validate_auto_accept_proof, ContactRequestDocumentParams,
+};
 use dpp::dashcore::secp256k1::rand::rngs::StdRng;
 use dpp::dashcore::secp256k1::rand::{RngCore, SeedableRng};
 use dpp::dashcore::secp256k1::{PublicKey, SecretKey};
 use dpp::data_contract::accessors::v0::DataContractV0Getters;
-use dpp::data_contract::document_type::accessors::DocumentTypeV0Getters;
 use dpp::document::DocumentV0;
 use dpp::identity::accessors::IdentityGettersV0;
 use dpp::identity::identity_public_key::accessors::v0::IdentityPublicKeyGettersV0;
@@ -22,134 +24,6 @@ use platform_encryption::{
     derive_shared_key_ecdh, encrypt_account_label, encrypt_extended_public_key, COMPACT_XPUB_LEN,
 };
 use std::collections::BTreeMap;
-
-use dpp::data_contract::DataContract;
-
-/// Already-derived crypto material and metadata for a DIP-15
-/// `contactRequest` document.
-///
-/// Everything here is plain data: the ECDH/encryption that produced
-/// `encrypted_public_key` and `encrypted_account_label`, and the randomness
-/// that produced `entropy`, happen in the caller.
-#[derive(Debug, Clone)]
-pub struct ContactRequestDocumentParams {
-    /// The sender's identity id (the document owner)
-    pub sender_id: Identifier,
-    /// The recipient's identity id (`toUserId`)
-    pub recipient_id: Identifier,
-    /// The sender's encryption key index used for ECDH
-    pub sender_key_index: u32,
-    /// The recipient's key index used for ECDH
-    pub recipient_key_index: u32,
-    /// Reference to the DashPay receiving account
-    pub account_reference: u32,
-    /// ECDH-encrypted extended public key: exactly 96 bytes
-    /// (16-byte IV + 80 bytes of encrypted DIP-15 compact xpub)
-    pub encrypted_public_key: Vec<u8>,
-    /// Optional encrypted account label: 48-80 bytes
-    /// (16-byte IV + 32-64 bytes of encrypted data)
-    pub encrypted_account_label: Option<Vec<u8>>,
-    /// Optional auto-accept proof (38-102 bytes) - not encrypted
-    pub auto_accept_proof: Option<Vec<u8>>,
-    /// The entropy that derives the document id; the same entropy must be
-    /// attached to the create transition, or platform consensus rejects it
-    /// with `InvalidDocumentTransitionIdError`.
-    pub entropy: [u8; 32],
-}
-
-/// Validate the size of a DIP-15 `autoAcceptProof` (38-102 bytes).
-pub fn validate_auto_accept_proof(proof: &[u8]) -> Result<(), Error> {
-    if proof.len() < 38 || proof.len() > 102 {
-        return Err(Error::Generic(format!(
-            "autoAcceptProof must be 38-102 bytes, got {}",
-            proof.len()
-        )));
-    }
-    Ok(())
-}
-
-/// Build the id and property map of a DIP-15 `contactRequest` document from
-/// already-derived crypto material.
-///
-/// This is the pure document-assembly half of [`Sdk::create_contact_request`]:
-/// the document id derives from `params.entropy`, and the property map
-/// carries exactly the fields the DashPay contract defines (`toUserId`,
-/// `encryptedPublicKey`, `senderKeyIndex`, `recipientKeyIndex`,
-/// `accountReference`, plus the optional `encryptedAccountLabel` and
-/// `autoAcceptProof`).
-///
-/// Returns `(document_id, properties)`.
-pub fn build_contact_request_document(
-    contract: &DataContract,
-    params: ContactRequestDocumentParams,
-) -> Result<(Identifier, BTreeMap<String, Value>), Error> {
-    if let Some(ref proof) = params.auto_accept_proof {
-        validate_auto_accept_proof(proof)?;
-    }
-
-    // Validate encrypted public key size (must be exactly 96 bytes: 16-byte IV + 80-byte encrypted data)
-    if params.encrypted_public_key.len() != 96 {
-        return Err(Error::Generic(format!(
-            "Encrypted public key size mismatch: expected 96 bytes, got {}",
-            params.encrypted_public_key.len()
-        )));
-    }
-
-    // Validate encrypted label size (48-80 bytes: 16-byte IV + 32-64 byte encrypted data)
-    if let Some(ref label) = params.encrypted_account_label {
-        if label.len() < 48 || label.len() > 80 {
-            return Err(Error::Generic(format!(
-                "Encrypted account label size out of range: expected 48-80 bytes, got {}",
-                label.len()
-            )));
-        }
-    }
-
-    let contact_request_document_type =
-        contract
-            .document_type_for_name("contactRequest")
-            .map_err(|_| {
-                Error::Generic("DashPay contactRequest document type not found".to_string())
-            })?;
-
-    let document_id = Document::generate_document_id_v0(
-        &contract.id(),
-        &params.sender_id,
-        contact_request_document_type.name(),
-        params.entropy.as_slice(),
-    );
-
-    let mut properties = BTreeMap::new();
-    properties.insert(
-        "toUserId".to_string(),
-        Value::Identifier(params.recipient_id.to_buffer()),
-    );
-    properties.insert(
-        "encryptedPublicKey".to_string(),
-        Value::Bytes(params.encrypted_public_key),
-    );
-    properties.insert(
-        "senderKeyIndex".to_string(),
-        Value::U32(params.sender_key_index),
-    );
-    properties.insert(
-        "recipientKeyIndex".to_string(),
-        Value::U32(params.recipient_key_index),
-    );
-    properties.insert(
-        "accountReference".to_string(),
-        Value::U32(params.account_reference),
-    );
-
-    if let Some(label) = params.encrypted_account_label {
-        properties.insert("encryptedAccountLabel".to_string(), Value::Bytes(label));
-    }
-    if let Some(proof) = params.auto_accept_proof {
-        properties.insert("autoAcceptProof".to_string(), Value::Bytes(proof));
-    }
-
-    Ok((document_id, properties))
-}
 
 /// ECDH provider for contact request encryption
 ///
@@ -387,7 +261,7 @@ impl Sdk {
         H: FnOnce(u32) -> Hut,
         Hut: std::future::Future<Output = Result<Vec<u8>, Error>>,
     {
-        // Validate auto accept proof size if provided. The builder
+        // Validate auto accept proof size if provided. The shared builder
         // validates again, but checking here first keeps the failure local —
         // before the recipient fetch and ECDH work below.
         if let Some(ref proof) = input.auto_accept_proof {
@@ -487,14 +361,14 @@ impl Sdk {
         let mut xpub_iv = [0u8; 16];
         rng.fill_bytes(&mut xpub_iv);
 
-        // Encrypt the extended public key (includes IV prepended). The
+        // Encrypt the extended public key (includes IV prepended). The shared
         // builder rejects any ciphertext that isn't exactly 96 bytes
         // (16-byte IV + 80-byte encrypted data).
         let encrypted_public_key =
             encrypt_extended_public_key(&shared_key, &xpub_iv, &extended_public_key);
 
         // Encrypt the account label if provided (includes IV prepended). The
-        // builder rejects any ciphertext outside 48-80 bytes
+        // shared builder rejects any ciphertext outside 48-80 bytes
         // (16-byte IV + 32-64 byte encrypted data).
         let encrypted_account_label = input.account_label.as_ref().map(|label| {
             let mut label_iv = [0u8; 16];
@@ -509,8 +383,8 @@ impl Sdk {
         let mut rng = StdRng::from_entropy();
         let entropy = Bytes32::random_with_rng(&mut rng);
 
-        // Assemble the document in the pure builder above, keeping document
-        // assembly separate from this networked flow.
+        // Assemble the document in the shared transport-free builder, so
+        // networked and embedder flows produce byte-identical documents.
         let sender_id = input.sender_identity.id().to_owned();
         let (document_id, properties) = build_contact_request_document(
             &dashpay_contract,
