+++ packages/rs-sdk/src/platform/dashpay/mod.rs
@@ -7,13 +7,14 @@ mod contact_request;
 mod contact_request_queries;
 
 pub use contact_request::{
-    build_contact_request_document, recipient_key_purpose_is_acceptable_on_receive,
-    recipient_key_purpose_is_valid, sender_key_purpose_is_acceptable_on_receive,
-    validate_auto_accept_proof, ContactRequestDocumentParams, ContactRequestInput,
-    ContactRequestResult, EcdhProvider, RecipientIdentity, SendContactRequestInput,
-    SendContactRequestResult,
+    recipient_key_purpose_is_acceptable_on_receive, recipient_key_purpose_is_valid,
+    sender_key_purpose_is_acceptable_on_receive, ContactRequestInput, ContactRequestResult,
+    EcdhProvider, RecipientIdentity, SendContactRequestInput, SendContactRequestResult,
 };
 pub use contact_request_queries::ContactRequestDocuments;
+pub use dash_platform_queries::dashpay::{
+    build_contact_request_document, validate_auto_accept_proof, ContactRequestDocumentParams,
+};
 
 use crate::platform::Fetch;
 use crate::{Error, Sdk};
